[ISSUE #9834]♻️Split the controller request processor into focused modules
